\name{G7}
\alias{G7}

\docType{data}

\title{Real Exchange Rates for the G7 Countries}

\description{
Real exchange rates for the G7 countries, for the years 1970--2003. 
}

\usage{
data("G7")
}

\format{
A data frame containing 204 observations on 6 variables.
\describe{
  \item{country}{A factor with levels \code{CAN}, \code{FRA}, \code{GBR}, \code{GER}, \code{ITA}, \code{JPN}.}
  \item{year}{Year.}
  \item{xrate}{Exchange rate wrt. the USA.}
  \item{ppp}{Purchasing power parity.}
  \item{rxrate}{Real exchange rate.}
  \item{lrxrate}{Real exchange rate, in logarithms.}
}
}

\details{
}

\source{
}

\references{

}


\examples{
data("G7")

## try: replicate Stata


}



\keyword{datasets}
