\name{OECDgdp}
\alias{OECDgdp}
\alias{OECDunemp}
\alias{OECDprod}
\docType{data}
\title{OECD statistics on GDP, unemployment rates, and industrial production indices.}
\description{
  These objects contain OECD seasonally adjusted quarterly data on GDP, harmonized 
  unemployment rates and industrial production indices for 8 countries 
  (Australia, Canada, France, Italy, Japan, Norway, United Kingdom and United States). 
  Data are from the public OECD web repository OECD.Stat Extracts.
}
\usage{
data("OECDgdp")
data("OECDunemp")
data("OECDprod")
}

\format{
  Data are are saved as binary data \code{.rda} objects. They are multiple time series objects.
}
\source{
    OECDStatExtracts.
}
\references{
    OECD, OECD.StatExtracts \url{http://stats.oecd.org/Index.aspx}.
}
\author{Claudio Lupi}
\keyword{datasets}
